Separate delete button from transcription; Added universal id `72` for allow delete message for all
