+++ src/handlers/markups/callbacks/mod.rs
@@ -1,3 +1,4 @@
 pub mod module;
 pub mod whisper;
 pub mod translate;
+pub mod delete;
