+++ src/util/transcription.rs
@@ -479,58 +479,3 @@ fn split_text(text: String, chunk_size: usize) -> Vec<String> {
         .map(|chunk| chunk.iter().collect())
         .collect()
 }
-
-pub async fn delete_transcription_handler(bot: Bot, query: CallbackQuery) -> Result<(), MyError> {
-    let Some(message_to_delete) = query.message else {
-        return Ok(());
-    };
-
-    let Some(message) = message_to_delete.regular_message() else {
-        return Ok(());
-    };
-
-    let Some(data) = query.data else {
-        return Ok(());
-    };
-    let clicker = query.from;
-
-    let target_user_id_str = data.strip_prefix("delete_").unwrap_or_default();
-    let Ok(target_user_id) = target_user_id_str.parse::<u64>() else {
-        bot.answer_callback_query(query.id)
-            .text("❌ Ошибка: неверный ID в кнопке.")
-            .show_alert(true)
-            .await?;
-        return Ok(());
-    };
-
-    let mut has_permission = false;
-
-    if clicker.id.0 == target_user_id {
-        has_permission = true;
-    }
-
-    if !has_permission && message.chat.is_group() || message.chat.is_supergroup() {
-        if let Ok(member) = bot.get_chat_member(message.chat.id, clicker.id).await {
-            if member.is_privileged() {
-                has_permission = true;
-            }
-        }
-    }
-
-    if !has_permission {
-        bot.answer_callback_query(query.id)
-            .text("❌ Удалить может только автор сообщения или администратор.")
-            .show_alert(true)
-            .await?;
-        return Ok(());
-    }
-
-    bot.answer_callback_query(query.id).await?;
-
-    bot.delete_message(message.chat.id, message.id)
-        .await
-        .map_err(|e| error!("Failed to delete bot's message: {:?}", e))
-        .ok();
-
-    Ok(())
-}
